requirements.txt
requests==2.26.0
Django~=4.1.7
djangorestframework==3.12.4
PyJWT==2.1.0
pytest~=7.2.2
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==23.1
py~=1.11.0
colorama~=0.4.6
cryptography~=39.0.2
pip~=23.0.1
toml~=0.10.2
tornado~=6.2
pytz~=2021.1
attrs~=22.2.0
tomli~=2.0.1
setuptools~=58.1.0
sqlparse~=0.4.3
asgiref~=3.6.0
Jinja2~=3.1.2
djoser~=2.1.0
MarkupSafe~=2.1.2
filters~=1.3.2
pluggy~=1.0.0
iniconfig~=2.0.0
atomicwrites~=1.4.1
certifi~=2022.12.7
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
uritemplate~=4.1.1
six~=1.16.0
oauthlib~=3.2.2
idna~=3.4
urllib3~=1.26.14
pycparser~=2.21
python-dotenv~=0.19.0